TestMoneroDaemonRpc non-relay tests compiling; defined MoneroWalletLocal difficulty long -> BigInteger
Former-commit-id: a873bbf34f6d050b0af7dfc35c8fcf1fa273425e

+++ src/main/java/monero/daemon/model/MoneroAltChain.java
@@ -1,5 +1,6 @@
 package monero.daemon.model;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -8,7 +9,7 @@ import java.util.List;
 public class MoneroAltChain {
   
   private List<String> blockIds;
-  private Long difficulty;
+  private BigInteger difficulty;
   private Integer height;
   private Integer length;
   private String mainChainParentBlockId;
@@ -21,11 +22,11 @@ public class MoneroAltChain {
     this.blockIds = blockIds;
   }
   
-  public Long getDifficulty() {
+  public BigInteger getDifficulty() {
     return difficulty;
   }
   
-  public void setDifficulty(Long difficulty) {
+  public void setDifficulty(BigInteger difficulty) {
     this.difficulty = difficulty;
   }
